% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_fof.R
\name{box_plot}
\alias{box_plot}
\title{Validation plot (MSEP/RMSEP/AUC)}
\usage{
box_plot(object, val_type = c("MSE", "RMSE"))
}
\arguments{
\item{object}{results from a cross-validation procedure.}

\item{val_type}{character.  Any of "RMSE", "MSE", or "AUC".}
}
\description{
Validation plot (MSEP/RMSEP/AUC)
}
\examples{
# 1D example:
\dontrun{

}
}
