% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_bases_fof_par.R
\name{cv_bases_fof_par}
\alias{cv_bases_fof_par}
\title{Cross-Validation for Functional Partial Least Squares with Basis Selection}
\usage{
cv_bases_fof_par(
  X,
  Y,
  center = TRUE,
  argvals_X = NULL,
  argvals_Y = NULL,
  num_bases_X = 20,
  num_bases_Y = 20,
  fda_basis_func_X = fda::create.bspline.basis,
  fda_basis_func_Y = fda::create.bspline.basis,
  penalty_X = 0,
  penalty_Y = 0,
  ncomp = min(10, ncol(X)),
  folds = 5,
  verbose = TRUE,
  stripped = TRUE,
  RPhi = NULL,
  RPsi = NULL,
  PX = NULL,
  PY = NULL,
  harmaccelLfd_X = NULL,
  harmaccelLfd_Y = NULL,
  ...
)
}
\arguments{
\item{X}{Matrix of predictors. Rows represent observations, and columns represent variables.}

\item{Y}{Matrix of responses. Rows represent observations, and columns represent variables.}

\item{center}{Logical; if \code{TRUE}, the data is mean-centered before fitting the model. Default is \code{TRUE}.}

\item{argvals_X}{Vector of argument values corresponding to the functional predictors. Default is \code{NULL}.}

\item{argvals_Y}{Vector of argument values corresponding to the functional responses. Default is \code{NULL}.}

\item{num_bases_X}{Integer or vector specifying the number of basis functions for the predictors. Default is 20.}

\item{num_bases_Y}{Integer or vector specifying the number of basis functions for the responses. Default is 20.}

\item{fda_basis_func_X}{Function to create the basis functions for the predictors. Defaults to \code{fda::create.bspline.basis}.}

\item{fda_basis_func_Y}{Function to create the basis functions for the responses. Defaults to \code{fda::create.bspline.basis}.}

\item{penalty_X}{Numeric penalty for smoothing the predictor basis functions. Default is 0.}

\item{penalty_Y}{Numeric penalty for smoothing the response basis functions. Default is 0.}

\item{ncomp}{Integer specifying the maximum number of components to compute. Default is \code{min(10, ncol(X))}.}

\item{folds}{Integer or list specifying the cross-validation folds. If an integer, it indicates the number of folds. If a list, it provides predefined fold indices. Default is 5.}

\item{verbose}{Logical; if \code{TRUE}, progress messages are displayed during cross-validation. Default is \code{TRUE}.}

\item{stripped}{Logical; if \code{TRUE}, the final model is not included in the output, only cross-validation results. Default is \code{TRUE}.}

\item{RPhi}{Penalty matrix for the predictors. Default is \code{NULL}.}

\item{RPsi}{Penalty matrix for the responses. Default is \code{NULL}.}

\item{PX}{Penalty matrix for the predictors. Default is \code{NULL}.}

\item{PY}{Penalty matrix for the responses. Default is \code{NULL}.}

\item{harmaccelLfd_X}{Object of class \code{fd} representing the harmonic acceleration operator for the predictors. Default is \code{NULL}.}

\item{harmaccelLfd_Y}{Object of class \code{fd} representing the harmonic acceleration operator for the responses. Default is \code{NULL}.}

\item{...}{Additional arguments passed to the \code{ffpls_bs} function.}
}
\value{
A list containing:
\describe{
\item{CVEs_ncomp}{Vector of cross-validation errors for each number of components.}
\item{MSE_ncomp_fold}{Matrix of mean squared errors (MSE) for each component and fold.}
\item{best_num_bases}{Matrix indicating the best number of basis functions for predictors and responses for each component.}
\item{final_model}{The final fitted FPLS model (only included if \code{stripped = FALSE}).}
\item{elapsed}{Elapsed time for the cross-validation process.}
}
}
\description{
Performs cross-validation to select the optimal number of basis functions for functional
predictors and responses in a functional partial least squares (FPLS) model.
}
\examples{
# 1D example:
}
