% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_unique_fof_par.R
\name{cv_unique_fof_par}
\alias{cv_unique_fof_par}
\title{Crossvalidation in parallel (same lambda for all components)}
\usage{
cv_unique_fof_par(
  X,
  Y,
  center = TRUE,
  argvals_X = NULL,
  argvals_Y = NULL,
  basisobj_X,
  basisobj_Y,
  penaltyvec_X = 0,
  penaltyvec_Y = 0,
  ncomp = min(10, ncol(X)),
  folds = 5,
  verbose = TRUE,
  stripped = TRUE,
  RPhi = NULL,
  RPsi = NULL,
  ...
)
}
\arguments{
\item{X}{a number-of-observations times nodes-of-X matrix.}

\item{Y}{a number-of-observations times nodes-of-Y matrix.}

\item{center}{logical, indicating if data should be centered.}

\item{argvals_X}{a set of argument values for X.}

\item{argvals_Y}{a set of argument values for Y.}

\item{basisobj_X}{basis object for X.}

\item{basisobj_Y}{basis object for Y.}

\item{penaltyvec_X}{a vector of possible penalties for the weights of X.}

\item{penaltyvec_Y}{a vector of possible penalties for the weights of Y.}

\item{ncomp}{number of components, integer.}

\item{folds}{a user defined list of folds (as generated by caret::createFolds())
or an integer indicating the number of folds.}

\item{verbose}{logical, indicating if messages should be printed.}

\item{stripped}{logical.  If \code{TRUE} the calculations are stripped as
much as possible for speed. Particularly, if \code{FALSE} (default) it computes
the final models using the best combination of penalties.
Inspired by package \code{pls}.}

\item{RPhi}{penalty matrix for the basis of X.}

\item{RPsi}{penalty matrix for the basis of Y.}

\item{...}{extra arguments for \code{pls::plsr()} function.}
}
\value{
A list of crossvalidates erros (IMSEs) and penalties giving the minimum
IMSEs per number of components.
}
\description{
Crossvalidation in parallel (same lambda for all components)
}
\examples{
# 1D example:
}
