% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffpls_bs.R
\name{ffpls_bs}
\alias{ffpls_bs}
\title{Penalized functional PLS based on the basis representation of the data}
\usage{
ffpls_bs(
  X,
  Y,
  center = TRUE,
  argvals_X,
  argvals_Y,
  ncomp = 3,
  basisobj_X,
  basisobj_Y,
  penalty_X = 0,
  penalty_Y = 0,
  verbose = TRUE,
  stripped = FALSE,
  RPhi = NULL,
  RPsi = NULL,
  ...
)
}
\arguments{
\item{X}{a number-of-observations times  nodes-of-X matrix.}

\item{Y}{a number-of-observations times nodes-of-Y matrix.}

\item{center}{logical, indicating if data should be centered.}

\item{argvals_X}{a set of argument values for X.}

\item{argvals_Y}{a set of argument values for Y.}

\item{ncomp}{number of components, integer.}

\item{basisobj_X}{basis object for X.}

\item{basisobj_Y}{basis object for Y.}

\item{penalty_X}{penalty for the weights of X.}

\item{penalty_Y}{penalty for the weights of Y.}

\item{verbose}{logical, indicating if messages should be printed.}

\item{stripped}{logical.  If \code{TRUE} the calculations are stripped as
much as possible for speed. Particularly, if \code{FALSE} (default) it computes
the final models using the best combination of penalties.
Inspired by package \code{pls}.}

\item{RPhi}{penalty matrix for the basis of X.}

\item{RPsi}{penalty matrix for the basis of Y.}

\item{...}{extra arguments for \code{pls::plsr()} function.}
}
\value{
a fof_pls model.
}
\description{
Penalized functional PLS based on the basis representation of the data
}
\examples{
# 1D example:
# library(fda)
}
