% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fofr_data.R
\name{generate_fofr_data}
\alias{generate_fofr_data}
\title{Generate artificial data for function-on-function regression}
\usage{
generate_fofr_data(
  nbasisX = 7,
  nbasisY = 5,
  nbeta = 1,
  nnodesX = 99,
  nnodesY = 98,
  Rsq = 0.9,
  n = 100,
  n_validation = 50
)
}
\arguments{
\item{nbasisX}{number of basis Bspline basis to use for X}

\item{nbasisY}{number of basis Bspline basis to use for Y}

\item{nbeta}{number identifying which beta(q, p) to generate}

\item{nnodesX}{number of nodes to use in the generation of X}

\item{nnodesY}{number of nodes to use in the generation of Y}

\item{Rsq}{R^2 between the generated Y(t) and perturbed Y(t) + error.}

\item{n}{sample size.}

\item{n_validation}{additional data for validation purposes.}
}
\value{
a list with the generate X, Y, and true beta
}
\description{
Generate artificial data for function-on-function regression
}
\examples{
generate_fofr_data(nbasisX = 7, nbasisY = 5, nbeta = 1,
                   nnodesX = 100, nnodesY = 100)
}
