% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_ffpls.R
\name{imse_beta_ffpls_bs}
\alias{imse_beta_ffpls_bs}
\title{Integrated mean square error for the evaluation of the estimated beta(q,p) FoF regression.}
\usage{
imse_beta_ffpls_bs(beta_true, beta_hat, argvals_X, argvals_Y)
}
\arguments{
\item{beta_true}{true beta(q, p)}

\item{beta_hat}{estimated beta(q, p)}

\item{argvals_X}{argument values "p" for X(p)}

\item{argvals_Y}{argument values "p" for X(q)}
}
\value{
a double
}
\description{
Integrated mean square error for the evaluation of the estimated beta(q,p) FoF regression.
}
\examples{
# Do not run
}
