% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_fof.R
\name{means_plot}
\alias{means_plot}
\title{Mean validation plot (mean of MSEP/RMSEP/AUC)}
\usage{
means_plot(object, val_type = c("MSE", "RMSE"))
}
\arguments{
\item{object}{results from a cross-validation procedure.}

\item{val_type}{character.  Any of "RMSE", "MSE", or "AUC".}
}
\description{
Mean validation plot (mean of MSEP/RMSEP/AUC)
}
\examples{
# 1D example:

\dontrun{
}
}
