% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_ffpls.R
\name{num_int_1d}
\alias{num_int_1d}
\title{1D numerical integration}
\usage{
num_int_1d(argvals, f_obs)
}
\arguments{
\item{argvals}{argument values of the function f(argvals)}

\item{f_obs}{function to integrate}
}
\value{
a double
}
\description{
1D numerical integration
}
\examples{
# Do not run
}
