% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_fof.R
\name{plot.cv_fofr}
\alias{plot.cv_fofr}
\title{Plot Methods for cv_sofr objects}
\usage{
\method{plot}{cv_fofr}(x, plot_type = c("means", "boxplot"), ...)
}
\arguments{
\item{x}{an object of class \code{cv_sofr} resulting from a cross-validation procedure.}

\item{plot_type}{character.  What kind of plot to plot.}

\item{...}{additional parameters that depends on the \code{plot_type}.}
}
\value{
returns whatever the underlying plot function.
}
\description{
Plot Methods for cv_sofr objects
}
\details{
The function is simply a wrapper for the underlying plot functions used to
make the selected plots.  See \code{box_plot} and \code{means_plot}
for details.
}
\examples{
# 1D example:
\dontrun{
}
}
