% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fofr_data.R
\name{generate_fofr_data}
\alias{generate_fofr_data}
\title{Generate artificial data for function-on-function regression}
\usage{
generate_fofr_data(
  nbasisX = 7,
  nbasisY = 5,
  nbeta = 1,
  nnodesX = 99,
  nnodesY = 98
)
}
\arguments{
\item{nbasisX}{number of basis Bspline basis to use for X}

\item{nbasisY}{number of basis Bspline basis to use for Y}

\item{nbeta}{number identifying which beta(q, p) to generate}

\item{nnodesX}{number of nodes to use in the generation of X}

\item{nnodesY}{number of nodes to use in the generation of Y}
}
\value{
a list with the generate X, Y, and true beta
}
\description{
Generate artificial data for function-on-function regression
}
\examples{
# To reproduce the paper by Preda & Schiltz (2011):
generate_fofr_data(nbasisX = 7, nbasisY = 5, nbeta = 1,
                   nnodesX = 100, nnodesY = 100)
}
